peru-github-users/requirements.txt
requests==2.31.0
python-dateutil==2.8.2
pandas==2.2.0
openai==1.12.0
python-dotenv==1.0.1
streamlit==1.31.0
plotly==5.18.0
altair==5.2.0
numpy==1.26.4
